fix: separate resolvers into different modules
