+++ src/resolvelib/__init__.py
@@ -23,5 +23,5 @@ from .resolvers import (
     ResolutionError,
     ResolutionImpossible,
     ResolutionTooDeep,
-    Resolver,
 )
+from .resolvers.resolution import Resolver
