+++ src/resolvelib/reporters.py
@@ -1,16 +1,11 @@
 from __future__ import annotations
 
-from typing import Generic
+from typing import TYPE_CHECKING, Collection, Generic
 
-from .structs import (
-    CT,
-    KT,
-    RT,
-    Collection,
-    Criterion,
-    RequirementInformation,
-    State,
-)
+from .structs import CT, KT, RT, RequirementInformation, State
+
+if TYPE_CHECKING:
+    from .resolvers.criterion import Criterion
 
 
 class BaseReporter(Generic[RT, CT, KT]):
