+++ tests/functional/cocoapods/test_resolvers_cocoapods.py
@@ -7,7 +7,8 @@ import string
 
 import pytest
 
-from resolvelib import AbstractProvider, ResolutionImpossible, Resolver
+from resolvelib import AbstractProvider, ResolutionImpossible
+from resolvelib.resolvers.resolution import Resolver
 
 Requirement = collections.namedtuple("Requirement", "name spec")
 Candidate = collections.namedtuple("Candidate", "name ver deps")
