+++ tests/functional/python/test_resolvers_python.py
@@ -11,7 +11,8 @@ import packaging.utils
 import packaging.version
 import pytest
 
-from resolvelib import AbstractProvider, ResolutionImpossible, Resolver
+from resolvelib import AbstractProvider, ResolutionImpossible
+from resolvelib.resolvers.resolution import Resolver
 
 Candidate = collections.namedtuple("Candidate", "name version extras")
